LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-select-element/customizable-select/select-picker-exit-animation-expected.txt


option

FAIL Top layer exit animations should work on ::picker(select) just like a popover. assert_equals: background-color should transition based on the exit animation. expected "rgb(128, 128, 128)" but got "rgb(240, 240, 240)"
